+++ src/main/java/org/truffleruby/language/loader/RequireNode.java
@@ -104,7 +104,6 @@ public abstract class RequireNode extends RubyBaseNode {
                 }
             }
 
-
             if (getContext().getOptions().LOG_AUTOLOAD && !toAutoload.isEmpty()) {
                 String info = toAutoload
                         .stream()
@@ -130,8 +129,11 @@ public abstract class RequireNode extends RubyBaseNode {
             final List<RubyConstant> releasedConstants = featureLoader.getAutoloadConstants(expandedPath);
             for (RubyConstant constant : releasedConstants) {
                 if (constant.getAutoloadConstant().isAutoloadingThread() && !alreadyAutoloading.contains(constant)) {
-                    final boolean undefined = GetConstantNode
-                            .autoloadUndefineConstantIfStillAutoload(constant);
+                    if (constant.getAutoloadConstant().shouldPublish()) {
+                        constant.getAutoloadConstant().publish(getContext(), constant);
+                    }
+
+                    final boolean undefined = GetConstantNode.autoloadUndefineConstantIfStillAutoload(constant);
                     GetConstantNode.logAutoloadResult(getContext(), constant, undefined);
                     GetConstantNode.autoloadConstantStop(constant);
                     featureLoader.removeAutoload(constant);